Cleanliness: unify the token scanners, extract encode.rs.
Two of the round-2 cleanliness asks:
- first_keyword and fenced_setting were near-identical hand-rolled token
  scanners (skip whitespace/comments, read a word). They now share one
  skip_trivia + next_word pair; next_word carries the double-quote handling,
  so both read "memory_limit" and memory_limit alike. ensure_single_statement
  keeps its own inline scan on purpose — it is a full-byte security lexer,
  not a tokenizer, and pinned by its own tests.
- the pure wire encoders (~690 lines: value/type emission, JSON-safe ints,
  temporal formatting, varint decimals, bit/uuid/base64, the KEYWORDS quoting
  table) move verbatim to encode.rs. Zero server state, so it is a mod + use
  shuffle: lib.rs drops 4560 -> 3869 lines and now reads as the server, with
  the encoders (tested against captured wire bytes) as their own subsystem.
  random_token stayed in lib.rs (it is a token helper, not an encoder).

All 10 suites green on both channels; behavior unchanged.
